requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
opencv-python==4.12.0.88
pillow==11.3.0
numpy==2.2.6
requests==2.32.3
python-dotenv==1.1.1
easyocr==1.7.2
python-multipart==0.0.20
scikit-image==0.25.2
scipy==1.16.1
torch==2.8.0
torchvision==0.23.0